+++ service/query_state.hh
@@ -13,7 +13,6 @@
 #include "service/client_state.hh"
 #include "tracing/tracing.hh"
 #include "service_permit.hh"
-#include "cql3/cql_statement.hh"
 
 namespace qos {
 class service_level_controller;
